feat: add opt-in Loki query cost guardrail for query_loki_logs (#1031)
Loki query cost is bytes scanned, determined only by the stream
selector and time range: line filters and parsers reduce what is
returned, not what is read, and Loki does not enforce
max_query_bytes_read on log queries without a line filter. LLM callers
routinely issue catch-all selectors over wide windows, which can scan
terabytes per call.

Gate query_loki_logs behind a three-mode guardrail (off by default;
shadow logs violations; enforce blocks): require a selective positive
label matcher in every stream selector, cap the effective time range
including range-vector durations, and pre-check Loki's index/stats
byte estimate against a budget, counting repeated selectors once per
occurrence with saturating arithmetic. Blocked queries return an MCP
tool error with rewrite guidance so models can retry cheaply;
unparseable queries fail open (Loki reports its own syntax errors) and
are logged. Scanning strips LogQL # comments and handles fractional
range-vector durations so commented-out or fractional syntax cannot
produce false blocks or bypasses. On VictoriaLogs the static checks
apply to selector-shaped queries and the byte pre-check is skipped,
since its stats approximation runs a real query.

Ref grafana/mcp-grafana#1030

+++ tools/loki_backend_victorialogs_test.go
@@ -177,7 +177,7 @@ func TestVictoriaLogsBackend_QueryStats_AppendsCountPipe(t *testing.T) {
 	// Other fields should remain zero — VL does not expose them.
 	assert.Equal(t, 0, stats.Streams)
 	assert.Equal(t, 0, stats.Chunks)
-	assert.Equal(t, 0, stats.Bytes)
+	assert.Equal(t, int64(0), stats.Bytes)
 
 	// Confirm we appended a stats pipe.
 	q := fake.lastForm.Get("query")
